Added clear data modal
