+++ src/html/header.html
@@ -86,7 +86,7 @@
         <a class="dropdown-item" href="#" id="menu_logout"><i class="fas fa-sign-out-alt"></i> <span data-i18n="logout">logout</span></a>
       </li>
       <li>
-        <a class="dropdown-item" href="#" id="menu_clear_data">
+        <a class="dropdown-item" href="#" data-toggle="modal" data-target="#clear_data_modal">
           <i class="fas fa-shield-alt"></i> <span data-i18n="clear_data">clear_data</span>
         </a>
       </li>
